Add setup rails config helper method
